fix: harden onboarding validation, JSON parsing, and test DB isolation
Code review of 8e24754 found three issues:

- validation.ts: budgetMax/minSizeSqm/renovationBudget map to Prisma
  Int columns but lacked .int(), so e.g. budgetMax: 3000000.5 was
  silently truncated instead of rejected. Added .int() to all three
  (minRooms stays a float — Prisma field is Float since room counts
  can be fractional, e.g. 3.5).
- route.ts: req.json() had no try/catch, so a malformed JSON body
  threw an uncaught SyntaxError -> opaque 500 instead of a clean 400.
- Integration tests were running against prisma/dev.db (the real dev
  database), and afterEach's deleteMany() wiped the whole table on
  every run. Added a dedicated prisma/test.db: .env.test holds its
  DATABASE_URL, tests/setup.ts loads it via dotenv as a vitest
  setupFile, and vitest.config.ts now references that setup file.
  dotenv was promoted from a transitive dependency to an explicit
  devDependency since test code now imports it directly.

  Note for later tasks/humans: the test DB schema is not created
  automatically — after a schema change, run:
    DATABASE_URL="file:./test.db" npx prisma migrate deploy
  from the project root to apply migrations to prisma/test.db.

Verified: inserted a sentinel row directly into dev.db, ran the full
test suite, and confirmed the sentinel row still existed afterward
(previously it would have been wiped by deleteMany()).

+++ src/lib/validation.ts
@@ -2,14 +2,14 @@ import { z } from "zod";
 
 export const onboardingSchema = z.object({
   locations: z.array(z.string()).min(1),
-  budgetMax: z.number().positive(),
+  budgetMax: z.number().int().positive(),
   minRooms: z.number().positive().optional(),
-  minSizeSqm: z.number().positive().optional(),
+  minSizeSqm: z.number().int().positive().optional(),
   mustHaveExtras: z.array(z.string()).default([]),
   goal: z.enum(["primary", "investment"]),
   openToRenting: z.boolean().default(false),
   openToFixerUpper: z.boolean().default(false),
-  renovationBudget: z.number().nonnegative().optional(),
+  renovationBudget: z.number().int().nonnegative().optional(),
   freeText: z.string().optional(),
   exampleUrls: z.array(z.string().url()).default([]),
 });
